gh-93649: Split heaptype tests from _testcapimodule.c
