+++ Modules/_testcapi/parts.h
@@ -1,3 +1,4 @@
 #include "Python.h"
 
 int _PyTestCapi_Init_Vectorcall(PyObject *module);
+int _PyTestCapi_Init_Heaptype(PyObject *module);
